@@ -2,10 +2,10 @@
 
 namespace Mako\CustomTableCard;
 
-use Laravel\Nova\Nova;
-use Laravel\Nova\Events\ServingNova;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
+use Laravel\Nova\Events\ServingNova;
+use Laravel\Nova\Nova;
 
 class CardServiceProvider extends ServiceProvider
 {
@@ -22,7 +22,7 @@ class CardServiceProvider extends ServiceProvider
 
         Nova::serving(function (ServingNova $event) {
             Nova::script('nova-custom-table-card', __DIR__.'/../dist/js/card.js');
-            Nova::style('nova-custom-table-card', __DIR__.'/../dist/css/card.css');
+            //Nova::style('nova-custom-table-card', __DIR__.'/../dist/css/card.css');
         });
     }
 
@@ -38,7 +38,7 @@ class CardServiceProvider extends ServiceProvider
         }
 
         Route::middleware(['nova'])
-                ->prefix('nova-vendor/customtablecard')
+                ->prefix('nova-vendor/nova-custom-table-card')
                 ->group(__DIR__.'/../routes/api.php');
     }
 
