@@ -1,4 +1,10 @@
 let mix = require('laravel-mix')
 
-mix.js('resources/js/card.js', 'dist/js')
-   .sass('resources/sass/card.scss', 'dist/css')
+require('./nova.mix')
+
+mix
+  .setPublicPath('dist')
+  .js('resources/js/card.js', 'js')
+  .vue({ version: 3 })
+  .css('resources/css/card.css', 'css')
+  .nova('m-a-k-o/nova-custom-table-card')
